@@ -16,16 +16,16 @@
         "issues": "https://github.com/middlewares/https/issues"
     },
     "require": {
-        "php": "^7.0",
-        "middlewares/utils": "^2.1",
+        "php": "^7.2",
+        "middlewares/utils": "^3.0",
         "psr/http-server-middleware": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.0|^7.0",
-        "zendframework/zend-diactoros": "^1.3",
+        "phpunit/phpunit": "^8.1",
+        "zendframework/zend-diactoros": "^2.2",
         "friendsofphp/php-cs-fixer": "^2.0",
         "squizlabs/php_codesniffer": "^3.0",
-        "phpstan/phpstan": "^0.9.2|^0.10.3"
+        "oscarotero/php-cs-fixer-config": "^1.0"
     },
     "autoload": {
         "psr-4": {
@@ -42,8 +42,7 @@
             "phpunit --coverage-text",
             "phpcs"
         ],
-        "cs-fix": "php-cs-fixer fix .",
-        "coverage": "phpunit --coverage-html=coverage",
-        "analyse": "phpstan analyse --no-progress --level 7 src"
+        "cs-fix": "php-cs-fixer fix",
+        "coverage": "phpunit --coverage-html=coverage"
     }
 }
